+++ backend/src/controllers/reports/index.js
@@ -319,6 +319,43 @@ controller.getProjectDashboardReportOnRequest = async (request, reply) => {
 };
 
 /**
+ * Get a Project Dashboard Report for a specific array of portfolio.
+ *
+ * @param   {FastifyRequest} request FastifyRequest is an instance of the standard http or http2 request objects.
+ * @param   {FastifyReply}   reply   FastifyReply is an instance of the standard http or http2 reply types.
+ * @returns {object}
+ */
+controller.getActiveProjectsReportOnRequest = async (request, reply) => {
+  controller.userRequires(request, what, "reports_read_all");
+  try {
+    const portfolios = request.query.portfolio;
+    const reportDate = new Date();
+    // Get the data from the database.
+    const result = {
+      active_projects: await model.getActiveProjects(portfolios),
+      report_date: reportDate.toLocaleDateString("en-US", {
+        day: "numeric",
+        month: "numeric",
+        year: "numeric",
+      }),
+    };
+    // todo: Uncomment when template document is created.
+    // const body = await getDocumentApiBody(result, "PA_StatusDashboard_template.docx");
+    // const pdf = await cdogs.api.post("/template/render", body, pdfConfig);
+    // request.data = pdf;
+    if (!result) {
+      reply.code(404);
+      return { message: `The ${what.single} with the specified id does not exist.` };
+    } else {
+      return result;
+    }
+  } catch (err) {
+    reply.code(500);
+    return { message: `There was a problem looking up this Active Projects Report.` };
+  }
+};
+
+/**
  * Separates an array of projects into groups by their portfolio.
  *
  * @param   {any[]}   rows Array of projects ordered by portfolio.